Update to C file

+++ 0x02-functions_nested_loops/101-natural.c
@@ -7,7 +7,7 @@
  * Return: 0
  */
 
-long int natural_35(void);
+int natural_35(void);
 
 int main(void)
 {
